services/masoc/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
kafka-python==2.0.2
pydantic==2.10.5
PyYAML==6.0.2
numpy==2.2.1
scikit-learn==1.6.0
requests==2.32.3
